QspiDxe: Qspi Driver Modifications
This commit does some changes in QSPI DXE driver
1. Break the Qspi Device tree parse function into three components. (as
per udit's comments)
2. Now that new device tree property uefi-runtime is introduced, use
that to check if the controller is going to be runtime or not?
3. Change the virtual address relocation method. earlier using link list
of controller data structures that are runtime and relocate those in
single event.
Now create separate event for each runtime controller.
4. Rework Write function so that TX watermark size (number of 4 byte
entries that can be filled at one time in TX fifo) is flexible instead
of fixed 4.

Signed-off-by: Pankaj Bansal <pankaj.bansal@nxp.com>