fdt_fixup: replace device tree loading library with driver
As we are moving towards the driver model, we expect that device
tree to be installed for UEFI drivers to use. The EmbeddedPkg
device tree driver relies on Variable service to be present.
Which in turn is provided by QSPI driver which needs device tree.

So, this is like chicken egg problem. to break this circle, we
define the device tree driver ourselves, which is based on EmbeddedPkg
driver but doesn't depends on runtime services to be present.

Signed-off-by: Pankaj Bansal <pankaj.bansal@nxp.com>