fdt_fixup: Add uefi-runtime property to Boot source
The boot source contains the Non Volatile variables.
Therefore, it is used by runtime services. Therefore it needs to
be hidden from OS.
Add a property uefi-runtime to device tree node, so that the DXE
driver corresponding to that device, is able to identify that the device
is supposed to be used at runtime.

Signed-off-by: Pankaj Bansal <pankaj.bansal@nxp.com>