i2c errata to improve clock
This errata is implemented in SOC code, Ideal place is driver.
There are two sort of boards
1) I2C first used by i2c driver
2) I2c first used to access fpga by means of i2c lib
We should apply errata before accessing the i2c lines.
If this errata to be applied in i2c specific code then
same code needs to go in lib and driver.
with this use-case in place, this errata is applied in SOC code

Signed-off-by: Udit Kumar <udit.kumar@nxp.com>