Silicon: NXP: Print SOC name and version
SOC name and version is needed for information purpose.
But these are not getting printed in release build.
It's because these were beng printed using DEBUG Macro with
DEBUG_INFO print level, which is not available in release build.
Therefore, change the print method from DEBUG Macro to
SerialPortWrite.

Contributed-under: TianoCore Contribution Agreement 1.1
Signed-off-by: Pankaj Bansal <pankaj.bansal@nxp.com>