Handle Root Bridge connected event
we need to fixup the IORT table in PciHostBridgeLib.
Therefore, delay the IORT table installation in AcpiPlatfromDxe.
Expose the IORT table pointer as a PCD so that other Drivers/Libs
can use it and after Root Bridge connected event is received, install
the IORT table.
By that time, IORT table fixups ought have to be done.

Contributed-under: TianoCore Contribution Agreement 1.1
Signed-off-by: Pankaj Bansal <pankaj.bansal@nxp.com>