Acpi/LX2160ARDB: remove nodes for BDFs other than 000
Since we will create the nodes for BDFs detected during PCIe enumeration
as part of IORT fixups, remove the nodes for BDFs other than 000.

BDF 000 is for PCIe controller itself which we can keep, because this is
bound to be present even if no other device is connected to PCIe
controller.

Contributed-under: TianoCore Contribution Agreement 1.1
Signed-off-by: Pankaj Bansal <pankaj.bansal@nxp.com>