Silicon/NXP: Modify I2cLib as per patches sent upstream
I2cLib current implementation is not correct, in that it treats all
the i2c transaction as reg read/write transactions. this is not correct,
as the i2c communication is dictated by peripheral and not by i2clib.

Therefore, modify the i2clib implementation to take care into acoount
the request packet passed on by peripheral.

Contributed-under: TianoCore Contribution Agreement 1.1
Signed-off-by: Pankaj Bansal <pankaj.bansal@nxp.com>