Silicon/NXP/I2cDxe: Fix hang issue with RTC
With latest edk2 codebase, sometimes hang is observed when Network
devices are being probed.
This is happening when an RTC read request is ongoing.
In between the read request, if an event occurs, which also triggers
RTC read, the I2c bus goes into unrecovered state.

This state is not even recovered, when rebooting the board.
We need to power off the board completely to recover i2c bus.

To prevent this, TPL level of I2c read is being raised to high, so that
no other even can preempt this. with this solution no hang has been
observed so far.

Signed-off-by: Pankaj Bansal <pankaj.bansal@nxp.com>