lx2: Add ACPI 6.0 introduced LPI(Low Power Idle).
LPI extensions leverages the processor container device(again introduced
in ACPI 6.0) allowing to express which parts of the system are affected
by a given LPI state. It defines the local power states for each node in
a hierarchical processor topology. The OSPM can use _LPI object to select
a local power state for each level of processor hierarchy in the system.

Signed-off-by: Peng Ma <peng.ma@nxp.com>