Platform/NXP: fixing register scaling in Ls1046aRDB and LS1046aFRWY
The ARM_GAS32 means spcr->BaseAddress.RegisterBitWidth == 32,
which means the registers are spaced a dword apart (dword aligned).
That's not actually true (DUART registers are byte-wide and follow
right after each other, with no padding), and it means the registers
are not correctly accessed. This is seen with ESXi.
Changed ARM_GAS32 with ARM_GAS8.

Signed-off-by: kuldip dwivedi <kuldip.dwivedi@puresoftware.com>