Silicon/NXP/I2cLib: Add Pcd for early I2cInit
The I2cDivisor is determined like this
I2cDivisor = (I2cClock + I2cSpeed - 1) / I2cSpeed
I2cClock is the input clock to I2c controller, I2cSpeed is the I2c bus
speed. In early init phase when we don't know the I2cClock, we need to
setup the I2c controller to be able to read SysClock (and hence I2cClock)
from the clock generator or FPGA connected to I2c bus.
Therefore, this Pcd selects the Divisor to use in early init phase.

Signed-off-by: Pankaj Bansal <pankaj.bansal@nxp.com>