PCIe : Move IO space towards starting of PCIe memory space
MMIO64 space should be aligned as per the size, to avoid
cases when EP needs MMIO64 space of large size.
We will create MMIO64 space of 16GB , so to align it with 16GB
boundary we will use the top 16GB space for MMIO64.
This requires move the IO space to front (region after the
config region space).

Signed-off-by: Wasim Khan <wasim.khan@nxp.com>