LS1046A : Fixed enumeration of MTD flash device on QSPI
_STA may return bit 0 clear (not present) with bit [3] set
indicating that the device is functional. This case is used
to indicate a valid device for which no device driver should
be loaded for example, a bridge device.
Children of this device may be present and valid. OSPM should
continue enumeration below a device whose _STA returns this bit
combination.

Signed-off-by: Vikas Singh <vikas.singh@puresoftware.com>