Silicon/NXP: Add DtInitDxe Driver
The NXP driver implementation uses the internal device-tree to
configure the drivers used by UEFI.  The DtInitDxe Driver allows
us to consume the DtPlatformLib and initialize the drivers without
the additional functionality of EmbeddedPkg's DtPlatformDxe such
as storing the boot method variable, which can't happen because
SPI Flash is not intialized yet.

I chose instead to make a smaller driver that loads the dtb and
sets an event handler for gEfiEventReadyToBootGuid that checks
DtAcpiPref variable and unsets gFdtTableGuid if device-tree is
not selected.

Signed-off-by: kuldip dwivedi <kuldip.dwivedi@puresoftware.com>