LS1046A : Update QBman S/W Portal regions in GCD Map.
SBSA Compliance v2.4 test 605 failed on LS1046ARDB and
LS1046AFRWY platforms.
This issue is mainly because “NonExist” memory entry (0000000580000000)
is not correct in the GCD map. The reason behind this wrong entry in map
is QBman sw portal regions which was mapped in-correctly in GCD map.
The total size of QBman sw portals region should be 256 MB
but it was wrongly mapped to 4GB.

Signed-off-by: Vikas Singh <vikas.singh@puresoftware.com>